# openapi.json
{"openapi":"3.1.0","info":{"title":"Soul spaces","description":"This is the API documentation for Soul spaces project.","version":"pre-v1"},"servers":[{"url":"http://localhost:8080","description":"Test"}],"paths":{"/api/v1/blogs/details":{"put":{"tags":["blog-controller"],"operationId":"updateBlogDetails","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/BlogUpdateRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlogUpdateResponse"}}}}}}},"/api/v1/blogs/access":{"put":{"tags":["blog-controller"],"operationId":"updateBlogAccess","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogAccessRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlogAccessResponse"}}}}}}},"/api/v1/blogs-details":{"put":{"tags":["blog-controller"],"operationId":"updatePost","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/BlogUpdateRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/{userId}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespone"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespone"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/login":{"post":{"tags":["authentication-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseBoolean"}}}}}}},"/auth/log-in":{"post":{"tags":["authentication-controller"],"operationId":"authenicate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthenticationRespone"}}}}}}},"/api/v1/upload":{"post":{"tags":["file-upload-controller"],"operationId":"uploadFile","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/uploadFiles":{"post":{"tags":["file-upload-controller"],"operationId":"uploadFiles","parameters":[{"name":"files","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"binary"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/blogs":{"post":{"tags":["blog-controller"],"operationId":"createBlog","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/BlogCreateRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlogCreateResponse"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getUList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUsuer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/auth/check-session":{"get":{"tags":["authentication-controller"],"operationId":"checkSession","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/v1/soulspaces":{"get":{"tags":["blog-controller"],"operationId":"getAllBlogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlogDTO"}}}}}}}},"/api/v1/blogs/{userId}/blogs":{"get":{"tags":["blog-controller"],"operationId":"getAllPublicBlogsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlogDTO"}}}}}}}},"/api/v1/blogs/{userId}/all":{"get":{"tags":["blog-controller"],"operationId":"getAllBlogsByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlogDTO"}}}}}}}},"/api/v1/blogs/{blogId}":{"get":{"tags":["blog-controller"],"operationId":"getBlogDetails","parameters":[{"name":"blogId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlogDetailsResponse"}}}}}},"delete":{"tags":["blog-controller"],"operationId":"deleteBlog","parameters":[{"name":"blogId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/delete":{"delete":{"tags":["file-upload-controller"],"operationId":"deleteFiles","parameters":[{"name":"imgUrl","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}}},"components":{"schemas":{"BlogUpdateRequest":{"type":"object","properties":{"blogId":{"type":"string"},"content":{"type":"string"},"newImages":{"type":"array","items":{"type":"string","format":"binary"}},"removeImagesUrl":{"type":"array","items":{"type":"string"}},"published":{"type":"boolean"}}},"BlogUpdateResponse":{"type":"object","properties":{"blogId":{"type":"string"},"content":{"type":"string"},"successUrls":{"type":"array","items":{"type":"string"}},"failedDeleteFiles":{"type":"array","items":{"type":"string"}},"failedUploadFiles":{"type":"array","items":{"type":"string"}},"published":{"type":"boolean"},"updateAt":{"type":"string","format":"date-time"}}},"BlogAccessRequest":{"type":"object","properties":{"blogId":{"type":"string"},"published":{"type":"boolean"}}},"BlogAccessResponse":{"type":"object","properties":{"blogId":{"type":"string"},"published":{"type":"boolean"},"updateAt":{"type":"string","format":"date-time"}}},"UserUpdateRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":2147483647,"minLength":3},"password":{"type":"string","maxLength":2147483647,"minLength":8},"fullname":{"type":"string"},"dob":{"type":"string","format":"date"}},"required":["password","username"]},"UserRespone":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"fullname":{"type":"string"},"dob":{"type":"string","format":"date"}}},"AuthenticationRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ApiResponseBoolean":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"type":"boolean"}}},"ApiResponseAuthenticationRespone":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/AuthenticationRespone"}}},"AuthenticationRespone":{"type":"object","properties":{"authentication":{"type":"boolean"}}},"BlogCreateRequest":{"type":"object","properties":{"userId":{"type":"string"},"content":{"type":"string"},"published":{"type":"boolean"},"files":{"type":"array","items":{"type":"string","format":"binary","description":"Danh sách file hình ảnh/blog"}}}},"BlogCreateResponse":{"type":"object","properties":{"blogId":{"type":"string"},"successUrls":{"type":"array","items":{"type":"string"}},"failedFiles":{"type":"array","items":{"type":"string"}}}},"UserCreationRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":2147483647,"minLength":3},"password":{"type":"string","maxLength":2147483647,"minLength":8},"fullname":{"type":"string"},"dob":{"type":"string","format":"date"}},"required":["password","username"]},"User":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"fullname":{"type":"string"},"dob":{"type":"string","format":"date"}}},"ApiResponseObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{}}},"BlogDTO":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"userName":{"type":"string"},"userAvatarUrl":{"type":"string"},"content":{"type":"string"},"imageContentUrls":{"type":"array","items":{"type":"string"}},"likeCount":{"type":"integer","format":"int64"},"commentCount":{"type":"integer","format":"int64"},"shareCount":{"type":"integer","format":"int64"},"published":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updateAt":{"type":"string","format":"date-time"}}},"BlogDetailsResponse":{"type":"object","properties":{"id":{"type":"string"},"userId":{"type":"string"},"userName":{"type":"string"},"userAvatarUrl":{"type":"string"},"content":{"type":"string"},"imageUrls":{"type":"array","items":{"type":"string"}},"likeCount":{"type":"integer","format":"int64"},"commentCount":{"type":"integer","format":"int64"},"shareCount":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"}}}}}}